requirements.txt
flask
flask-cors
gunicorn
python-dotenv

# Firebase
firebase-admin

# Auth
bcrypt

# Optional (already used sometimes)
requests